backend_catshop/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
python-dotenv==1.0.0
firebase-admin==6.3.0
cloudinary==1.32.0

pydantic>=2.0
pydantic-settings
PyJWT
Pillow>=10.0.0
opencv-python-headless
numpy

ultralytics==8.1.0
